Fixup a few android contacts things
In these 2 cases Mimetype is added to the query but was not checking to
see if
anything was already there.

This will avoid queries where the AND is missing

((mimetype = ?) AND (data2 IS NOT NULL))(mimetype = ?) bad
((mimetype = ?) AND (data2 IS NOT NULL)) AND (mimetype = ?) better

It is not clear why Android is not returning some of the ids that
are requested.  This fix will simply skip over any missing ids.

Skipping the ids should match the behavior before this commit